feat: Instead of duplicating the entire connection API, have a type parameter (#3619)
## Description

on Connection that describes the connection state, and implement the fns
alpn and remote_id differently depending on what the state is.

The upside is that it should be easier to write code that works with
both a 0rtt connection and a fully initialized connection without having
to define a trait to abstract over both variants.

The downside is that it is a bit fancy with the types. But then for the
non 0rtt case you would never see the fanciness, so maybe it is OK?

Todo:
- remove everything but connection and handshake_completed from
Incoming/Outcoming...
- make the ConnectionState trait sealed
- document the ConnectionState instances

Note:

if you have a handler that needs to work with both handshake completed
connections and non handshake completed connections, you could now just
do:

```rust
async fn handler<T: ConnectionState>(conn: Connection<T>) { ... }
```

## Breaking Changes

- `iroh::endpoint::Connection` gets a type parameter to track connection
state. Default is `Connection<HandshakeCompleted>`
- `iroh::endpoint::IncomingZeroRttConnection` is removed, replaced with
type alias to `Connection<IncomingZeroRtt>`
- `iroh::endpoint::OutgoingZeroRttConnection` is removed, replaced with
type alias to `Connection<OutgoingZeroRtt>`

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [x] All breaking changes documented.
- [ ] List all breaking changes in the above "Breaking Changes" section.
- [ ] Open an issue or PR on any number0 repos that are affected by this
breaking change. Give guidance on how the updates should be handled or
do the actual updates themselves. The major ones are:
    - [ ] [`quic-rpc`](https://github.com/n0-computer/quic-rpc)
    - [ ] [`iroh-gossip`](https://github.com/n0-computer/iroh-gossip)
    - [ ] [`iroh-blobs`](https://github.com/n0-computer/iroh-blobs)
    - [ ] [`dumbpipe`](https://github.com/n0-computer/dumbpipe)
    - [ ] [`sendme`](https://github.com/n0-computer/sendme)

---------

Co-authored-by: Friedel Ziegelmayer <me@dignifiedquire.com>
Co-authored-by: Frando <frando@unbiskant.org>