refactor(multipath): move discovery into `EndpointStateActor` (#3645)
## Description

Fixes #3642 

This moves discovery handling fully into the `EndpointStateActor`.
The pub(crate) interface to trigger discovery and get a
EndpointMappedAddr is now `Magicsock::resolve_remote`, which sends the
provided addresses to the EndpointStateActor. The actor starts discovery
if it does not have a selected path and if discovery is not running. It
returns either immediately if there are any known paths, or waits for
discovery to produce at least one result or an error. Once this returns,
`resolve_remote` returns either with a EndpointMappedAddr or with the
discovery error.

This means the current behavior is kept: We only start
`quinn::Endpoint::connect` once we have at least one transport address
for the remote. If not, we return the discovery error immediately from
`iroh::Endpoint::connect`.

This opens the door for us to easily tune when to run discovery in other
siutations, e.g. when all available paths to a remote are closed.
However, for now this PR still only starts discovery when
`Endpoint::connect` is called and no path is selected at the moment.


## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist
<!-- Remove any that are not relevant. -->
- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.
- [ ] List all breaking changes in the above "Breaking Changes" section.
- [ ] Open an issue or PR on any number0 repos that are affected by this
breaking change. Give guidance on how the updates should be handled or
do the actual updates themselves. The major ones are:
    - [ ] [`quic-rpc`](https://github.com/n0-computer/quic-rpc)
    - [ ] [`iroh-gossip`](https://github.com/n0-computer/iroh-gossip)
    - [ ] [`iroh-blobs`](https://github.com/n0-computer/iroh-blobs)
    - [ ] [`dumbpipe`](https://github.com/n0-computer/dumbpipe)
    - [ ] [`sendme`](https://github.com/n0-computer/sendme)