feat(iroh): introduce EndpointHooks (#3688)
## Description

This adds the conecpt of hooks to the iroh endpoint. `Hooks` are structs
implementing the `EndpointHooks` trait and are used to intercept the
establishment of connections. Multiple hooks can be added to the
endpoint, and they will be invoked for each hook in the order they have
been added to the endpoint.

Currently there's two methods on the `EndpointHooks` trait:

* `before_connect` is invoked before an outgoing connection is started.
* `after_handshake` is invoked for incoming and outgoing connections
once the TLS handshake has completed

Both methods return an `Outcome`, which can either be `Reject` or
`Accept`. If any hook returns `Reject`, the connection or connection
attempt will be rejected.

The PR also adds `ConnectionInfo`, which is a struct that has
information about a connection, but does not keep the connection itself
alive. It allows to inspect stats and paths, and there's a `closed`
method that returns a future which completes once the connection closes
(without keeping the connection alive).

The PR includes two examples:
* `auth-hook` implements authentication for iroh protocols through a
middleware and a separate authentication protocol. Individual protocols
don't need to be aware of authentication at all.
* `monitor-connnections` monitors incoming and outgoing connections and
prints connection stats once a connection closes.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist
<!-- Remove any that are not relevant. -->
- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.
- [ ] List all breaking changes in the above "Breaking Changes" section.
- [ ] Open an issue or PR on any number0 repos that are affected by this
breaking change. Give guidance on how the updates should be handled or
do the actual updates themselves. The major ones are:
    - [ ] [`quic-rpc`](https://github.com/n0-computer/quic-rpc)
    - [ ] [`iroh-gossip`](https://github.com/n0-computer/iroh-gossip)
    - [ ] [`iroh-blobs`](https://github.com/n0-computer/iroh-blobs)
    - [ ] [`dumbpipe`](https://github.com/n0-computer/dumbpipe)
    - [ ] [`sendme`](https://github.com/n0-computer/sendme)

---------

Co-authored-by: dignifiedquire <me@dignifiedquire.com>
Co-authored-by: ramfox <kasey@n0.computer>