docs(iroh-relay): add section on how to use as a library (#3715)
## Description

Add a section to the `iroh-relay` README that describes how to run an
in-process relay server when using iroh/iroh-relay as a library.

## Breaking Changes

## Notes & open questions

I was creating an instance of the relay until finding out about the
function in `test-utils`, and skipping the certificate verification was
hard to find as well.

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.

Co-authored-by: Philipp Krüger <philipp.krueger1@gmail.com>