fix(iroh): Ensure the selected holepunched path is set to `PathStatus::Available` (#3771)
## Description

This should fix a possible bug that works like this:
- Connection is opened using a MixedMappedAddr, and the relay wins, so
we end up with only a single relay addr.
- The client side initiates holepunching.
- The client side's quinn opens a bunch of paths. For that it uses
`open_path_ensure` inside quinn with `PathStatus::Backup` (This is good!
I initially wanted to fix that, but ever since QNT this is already
fixed!)
- Once the first IP path holepunches, the RemoteStateActor on the client
side will see the path open. It'll `select_path` and call
`open_path_ensure` on all connections it has with the best path's remote
addr and `PathStatus::Available`.
- However, because QNT already opened a path there, `open_path_ensure`
doesn't do anything, thus doesn't switch the path to available. We're
now stuck with only backup paths, so **keep sending on the relay**.


## Notes & open questions

Someone please tell me how to best test/verify this. Perhaps I should
log all the known paths and their statuses on the connection? Or
something like that?

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.