feat: expose known remote addrs (#3752)
## Description

Adds `Endpoint::remote_info` which returns `Option<RemoteInfo>`. The
`RemoteInfo` exposes an iterator over all transport addrs for this
remote that we know about.

Fixes #3521

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

This basically copies the `RemotePathInfo` from the RemoteStateActor
into a new struct `RemoteInfo`. So, this an async fn over the channel to
the RemoteStateActor, and allocates. As this is a rather niche API, I
think that's fine.

I wasn't entirely sure how much info we want to expose. Right now, the
PR only includes the latest `Source` for the address, as I wasn't sure
if we really want or need to expose the full list of sources.

## Change checklist
<!-- Remove any that are not relevant. -->
- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.
- [ ] List all breaking changes in the above "Breaking Changes" section.
- [ ] Open an issue or PR on any number0 repos that are affected by this
breaking change. Give guidance on how the updates should be handled or
do the actual updates themselves. The major ones are:
    - [ ] [`quic-rpc`](https://github.com/n0-computer/quic-rpc)
    - [ ] [`iroh-gossip`](https://github.com/n0-computer/iroh-gossip)
    - [ ] [`iroh-blobs`](https://github.com/n0-computer/iroh-blobs)
    - [ ] [`dumbpipe`](https://github.com/n0-computer/dumbpipe)
    - [ ] [`sendme`](https://github.com/n0-computer/sendme)