fix(deps): Update to newest quinn main and fix api usage (#3802)
## Description

The `quinn_proto::coding::Codec` API changed to be split in `Encodable`
and `Decodable`. That broke the minimal versions checks somehow.

Fixed by updating things to use the latest quinn from main and adjusting
to the new API.

## Breaking Changes

Technically this is exported on the public API for us. Though I think
that's an error. Plus I don't think this has been published yet.

## Notes & open questions

I'll file a separate PR to remove these from the public API.

## Change checklist

- [x] Self-review.
- [x] All breaking changes documented.
- [x] List all breaking changes in the above "Breaking Changes" section.